+++ lib/src/helpers/params.dart
@@ -1,0 +1,7 @@
+extension Params on String {
+  /// Get the param as int
+  int toInt() => int.parse(toString());
+
+  /// Get the param as double
+  double toDouble() => double.parse(toString());
+}